@@ -1,25 +1,28 @@
 import 'dart:convert';
 import 'dart:typed_data';
 
-import 'package:migrator/common/crypto.dart';
+import 'package:migrator/utils/crypto.dart';
 
 class Connection {
+  String id;
   String name;
   String host;
-  String username;
-  String password;
-  Uint8List certificate;
+  String? username;
+  String? password;
+  Uint8List? certificate;
 
   Connection({
-    this.name,
-    this.host,
+    required this.id,
+    required this.name,
+    required this.host,
     this.username,
     this.password,
     this.certificate,
   });
 
   Connection.fromJson(Map<String, dynamic> json)
-      : name = json['name'],
+      : id = json['id'] ?? '',
+        name = json['name'],
         host = json['host'],
         username = json['username'],
         password = decrypt(json['password'], throwError: false),
@@ -28,11 +31,13 @@ class Connection {
             : null;
 
   Map<String, dynamic> toJson() => {
+        'id': id,
         'name': name,
         'host': host,
         'username': username,
-        'password': encrypt(password, throwError: false),
-        'certificate': certificate != null ? base64Encode(certificate) : null,
+        'password':
+            password != null ? encrypt(password!, throwError: false) : null,
+        'certificate': certificate != null ? base64Encode(certificate!) : null,
       };
 
   @override
