@@ -2,28 +2,32 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'utils.dart';
-
 Future<HttpClientResponse> http(
   String method,
   String url, {
-  Map<String, dynamic> params,
-  Map<String, String> headers,
-  String body,
-  String contentType,
-  String username,
-  String password,
-  Uint8List certificate,
-  Function(HttpClientRequest) requestHook,
+  Map<String, dynamic?> params = const {},
+  Map<String, String?> headers = const {},
+  String? body,
+  String? contentType,
+  String? username,
+  String? password,
+  Uint8List? certificate,
+  Function(HttpClientRequest)? requestHook,
 }) async {
-  String basicAuth;
-  SecurityContext securityCtx;
+  String? basicAuth;
+  SecurityContext? securityCtx;
   HttpClientRequest request;
 
   try {
-    if (certificate == null) {
+    if (username != null &&
+        password != null &&
+        username.isNotEmpty &&
+        password.isNotEmpty) {
       basicAuth = 'Basic ' + base64Encode(utf8.encode('$username:$password'));
-    } else {
+    } else if (certificate != null &&
+        password != null &&
+        certificate.isNotEmpty &&
+        password.isNotEmpty) {
       securityCtx = SecurityContext(withTrustedRoots: true);
       securityCtx.useCertificateChainBytes(certificate, password: password);
       securityCtx.usePrivateKeyBytes(certificate, password: password);
@@ -31,7 +35,7 @@ Future<HttpClientResponse> http(
 
     var uri = Uri.parse(url);
 
-    if (params != null && params.isNotEmpty) {
+    if (params.isNotEmpty) {
       params.removeWhere((key, value) => value == null);
       uri = Uri(
         scheme: uri.scheme,
@@ -46,7 +50,7 @@ Future<HttpClientResponse> http(
 
     request = await http.openUrl(method, uri);
 
-    if (headers != null && headers.isNotEmpty) {
+    if (headers.isNotEmpty) {
       headers.forEach((key, value) {
         if (value != null) {
           request.headers.add(key, value);
@@ -74,6 +78,6 @@ Future<HttpClientResponse> http(
 
     return response;
   } on SocketException {
-    throw Failure('No Internet connection 😑');
+    throw Exception('Revisa la conexión a internet 😑');
   }
 }
