@@ -1,16 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:get_it/get_it.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'package:migrator/app.dart';
-import 'package:migrator/services/restman_service.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import 'app.dart';
 
 void main() {
-  setup();
-  runApp(App());
-}
-
-void setup() {
-  GetIt.I.registerSingleton<RestmanService>(RestmanService(useCache: false));
-  GetIt.I.registerSingletonAsync(() => SharedPreferences.getInstance());
+  runApp(ProviderScope(child: App()));
 }
