@@ -1,174 +1,212 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:styled_widget/styled_widget.dart';
 
-import 'package:migrator/common/common.dart';
+import 'package:migrator/utils/utils.dart';
 import 'package:migrator/widgets/widgets.dart';
 import 'package:migrator/models/models.dart';
-import 'package:migrator/stores/stores.dart';
+import 'package:migrator/providers/providers.dart';
 
 import 'connections_screen.dart';
 import 'select_items_screen.dart';
 
-class SelectConnectionsScreen extends StatefulWidget {
-  @override
-  _SelectConnectionsScreenState createState() =>
-      _SelectConnectionsScreenState();
-}
+class SelectConnectionsScreen extends HookWidget {
+  Future<void> _continue(BuildContext context) async {
+    final source = context.read(sourceConnectionProvider);
+    final target = context.read(targetConnectionProvider);
 
-class _SelectConnectionsScreenState extends State<SelectConnectionsScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Theme.of(context).backgroundColor,
-      appBar: AppBar(
-        title: const Text('Selecionar Conexiones'),
-        actions: [
-          ActionButton(
-            icon: Icons.arrow_forward_outlined,
-            label: 'Continuar',
-            onPressed: () {
-              final store = context.store<SelectedConnectionsStore>();
-              if (store.sourceConnection == null ||
-                  store.targetConnection == null) {
-                alert(
-                  context,
-                  title: Text('Falto algo...'),
-                  content: Text(
-                    'Debes selecionar una conexión origen y otra destino para continuar',
-                  ),
-                );
-                return;
-              }
-              push(context, (_) => SelectItemsScreen());
-            },
-          ),
-          ActionButton(
-            icon: Icons.public_outlined,
-            label: 'Conexiones',
-            onPressed: () => push(context, (_) => ConnectionsScreen()),
-          )
+    final ok = await confirm(
+      context,
+      title: Text('¿Continuar con esta selección?'),
+      content: Wrap(
+        direction: Axis.vertical,
+        spacing: 10.0,
+        children: [
+          Text('Origen: ${source.state}'),
+          Text('Destino: ${target.state}'),
         ],
       ),
+    );
+
+    if (!ok) {
+      source.state = null;
+      return;
+    }
+
+    push(context, (_) => SelectItemsScreen());
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    useEffect(() {
+      Future.microtask(() {
+        context.read(sourceConnectionProvider).state = null;
+        context.read(targetConnectionProvider).state = null;
+      });
+    }, []);
+    return Scaffold(
+      backgroundColor: Theme.of(context).backgroundColor,
+      appBar: _buildAppBar(),
       body: _buildBody(),
     );
   }
 
-  Widget _buildBody() {
-    final store = context.store<ConnectionsStore>();
+  AppBar _buildAppBar() {
+    final context = useContext();
 
-    if (store.connections.length == 0) {
-      return Text('No hay conexiones configuradas')
-          .textColor(Colors.white)
-          .center();
-    }
-
-    if (store.connections.length == 1) {
-      return Text('Solo hay una conexión, registrar otra')
-          .textColor(Colors.white)
-          .center();
-    }
-
-    return Column(
-      children: [
-        _buildBodyContent(),
-        _buildStatusBar(),
+    return AppBar(
+      title: const Text('Selecionar Conexiones'),
+      actions: [
+        ActionButton(
+          icon: Icons.public_outlined,
+          label: 'Conexiones',
+          onPressed: () => push(context, (_) => ConnectionsScreen()),
+        ),
+        const SizedBox(width: 5.0),
       ],
     );
   }
 
-  Widget _buildBodyContent() {
-    return Row(
-      children: [
-        _buildPanel(true),
-        _buildPanel(false),
-      ],
-    ).flexible(fit: FlexFit.tight);
+  Widget _buildBody() {
+    final context = useContext();
+    final asyncList = useProvider(connectionListProvider.state);
+
+    useEffect(() {
+      Future.microtask(() => context.read(connectionListProvider).fetch());
+    }, []);
+
+    return asyncList.when(
+      data: (connections) {
+        if (connections.length == 0) {
+          return _buildNeedConnectionFeedback('No hay conexiones configuradas');
+        }
+
+        if (connections.length == 1) {
+          return _buildNeedConnectionFeedback(
+            'Solo hay una conexión, debes registrar otra',
+          );
+        }
+
+        return Row(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            _buildPanel(true),
+            VerticalDivider(),
+            _buildPanel(false),
+          ],
+        );
+      },
+      loading: () =>
+          Text('Cargando las conexiones...').textColor(Colors.white).center(),
+      error: (e, st) => Text('Error cargando las conexiones: ${e.toString()}')
+          .textColor(Colors.white)
+          .center(),
+    );
   }
 
-  Widget _buildStatusBar() {
-    final store = context.store<SelectedConnectionsStore>();
-    return StatusBar(
-      child: Observer(builder: (_) {
-        var message;
+  Widget _buildNeedConnectionFeedback(String message) {
+    final context = useContext();
 
-        if (store.targetConnection != null) {
-          message = 'Conexiones selecionadas, puedes continuar';
-        } else if (store.sourceConnection != null) {
-          message = 'Seleciona la conexión destino';
-        }
-
-        if (message == null) {
-          return SizedBox();
-        }
-
-        return Indicator(
-          Text(message),
-          icon: Icons.info_rounded,
-          color: Colors.blue,
-          size: 16.0,
-        );
-      }),
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(message).fontSize(22.0).textColor(Colors.white),
+          SizedBox(height: 8.0),
+          ElevatedButton(
+            style: ElevatedButton.styleFrom(
+              primary: Theme.of(context).primaryColor,
+              textStyle: TextStyle(fontSize: 18.0),
+            ),
+            child: Text('Crear conexión'),
+            onPressed: () => push(context, (_) => ConnectionsScreen()),
+          )
+        ],
+      ),
     );
   }
 
   Widget _buildPanel(bool isSource) {
-    final title = isSource ? 'Origen' : 'Destino';
-    return Column(
-      children: [
-        Text(title).fontWeight(FontWeight.bold).padding(top: 20.0),
-        _buildList(isSource),
-      ],
-    ).card(elevation: 8).expanded();
-  }
-
-  Widget _buildList(bool isSource) {
-    final selectedConnectionsStore = context.store<SelectedConnectionsStore>();
-    final connectionsStore = context.store<ConnectionsStore>();
-    return Observer(
-      builder: (context) {
-        final connections = connectionsStore.connections.where((conn) {
-          return isSource
-              ? true
-              : (conn != selectedConnectionsStore.sourceConnection);
-        }).toList();
-
-        if (!isSource && selectedConnectionsStore.sourceConnection == null) {
-          return Text('Selecionar la conexión origen').center().expanded();
-        }
-
-        return ListView.separated(
-          itemCount: connections.length,
-          shrinkWrap: true,
-          separatorBuilder: (_, __) => Divider(
-            color: Theme.of(context).primaryColor,
-          ),
-          itemBuilder: (_, index) => _buildTile(isSource, connections[index]),
-        );
-      },
+    return Expanded(
+      child: Container(
+        color: Colors.white,
+        child: Column(
+          children: [
+            _buildPanelTop(isSource),
+            _buildPanelList(isSource).expanded(),
+          ],
+        ),
+      ),
     );
   }
 
-  Widget _buildTile(bool isSource, Connection connection) {
-    final store = context.store<SelectedConnectionsStore>();
-    return Observer(builder: (_) {
-      final current =
-          isSource ? store.sourceConnection : store.targetConnection;
-      final selected = current == connection;
-      return ListTile(
-        leading: Icon(Icons.public).padding(right: 12.0).border(right: 1.0),
-        title: Text(connection.toString()),
-        trailing: selected ? Icon(Icons.check_box) : null,
-        selected: selected,
-        onTap: () {
-          if (isSource) {
-            store.sourceConnection = connection;
-            store.targetConnection = null;
-          } else {
-            store.targetConnection = connection;
-          }
-        },
-      );
-    });
+  Widget _buildPanelTop(bool isSource) {
+    final context = useContext();
+    final sourceConnection = useProvider(sourceConnectionProvider).state;
+    final targetConnection = useProvider(targetConnectionProvider).state;
+
+    String message = '';
+
+    if (isSource && sourceConnection == null) {
+      message = 'Selecionar ambiente origen';
+    }
+
+    if (!isSource && sourceConnection != null && targetConnection == null) {
+      message = 'Selecionar ambiente destino';
+    }
+
+    return Text(message)
+        .fontSize(18.0)
+        .textColor(Theme.of(context).primaryColor)
+        .fontWeight(FontWeight.w500)
+        .padding(vertical: 10.0);
+  }
+
+  Widget _buildPanelList(bool isSource) {
+    final asyncList = useProvider(connectionListFamily(isSource));
+
+    return asyncList.maybeWhen(
+      data: (connections) => ListView.separated(
+        itemCount: connections.length,
+        shrinkWrap: true,
+        separatorBuilder: (context, __) => Divider(
+          color: Theme.of(context).primaryColor,
+        ),
+        itemBuilder: (context, index) => _buildTile(
+          context,
+          isSource,
+          connections[index],
+        ),
+      ),
+      orElse: () => Text('!!Upps this should not happen.'),
+    );
+  }
+
+  Widget _buildTile(
+    BuildContext context,
+    bool isSource,
+    Connection connection,
+  ) {
+    final sourceConnection = context.read(sourceConnectionProvider);
+    final targetConnection = context.read(targetConnectionProvider);
+    final current = isSource ? sourceConnection.state : targetConnection.state;
+    final isSelected = current?.id == connection.id;
+
+    return ListTile(
+      leading: Icon(Icons.public).padding(right: 12.0).border(right: 1.0),
+      title: Text(connection.toString()),
+      trailing: isSelected ? Icon(Icons.check_box) : null,
+      selected: isSelected,
+      onTap: () {
+        if (isSource) {
+          sourceConnection.state = connection;
+          targetConnection.state = null;
+        } else {
+          targetConnection.state = connection;
+          _continue(context);
+        }
+      },
+    );
   }
 }
