@@ -5,7 +5,7 @@ final key = Key.fromUtf8(pwd);
 final iv = IV.fromLength(16);
 final encrypter = Encrypter(AES(key));
 
-String encrypt(String plainText, {throwError: true}) {
+String? encrypt(String plainText, {throwError: true}) {
   try {
     final encrypted = encrypter.encrypt(plainText, iv: iv);
 
@@ -19,7 +19,7 @@ String encrypt(String plainText, {throwError: true}) {
   return null;
 }
 
-String decrypt(String encoded, {throwError: true}) {
+String? decrypt(String encoded, {throwError: true}) {
   try {
     final decrypted = encrypter.decrypt64(encoded, iv: iv);
 
