@@ -1,5 +1,6 @@
 export 'action_button.dart';
 export 'form_builder_file_field.dart';
-export 'indicator.dart';
 export 'migrate_screen_body.dart';
-export 'statusbar.dart';
+export 'node_icon.dart';
+export 'selected_connections_bar.dart';
+export 'split_view.dart';
